cgp: added geometric length distribution of implicit introns, that penalizes long introns.
git-svn-id: file:///home/gumz3/Desktop/svnAugustusRepo/svnArchiveFile/svn/augustus/trunks@1066 946bced8-b911-4d30-a262-e0a769a8c5ad

Former-commit-id: b4110c781b9dd4daf374524f0d2cd43ebf0de601

+++ src/orthograph.cc
@@ -365,7 +365,7 @@ double OrthoGraph::dualdecomp(ExonEvo &evo, vector< list<Transcript*> *> &geneli
 			bool h = node->label;
 			bool v = hects->labels[pos];
 			if(v != h){  //shared nodes are labelled inconsistently in the two subproblems
-			    double weight = delta*(v-h);
+			    float weight = delta*(v-h);
 			    //update weights
 			    node->addWeight(weight);
 			    hects->weights[pos] -= weight;
